Move eval label exn, rename util to eval_tags
