+++ tests/aisystems/test_ai_systems_eval_tags.py
@@ -1,4 +1,4 @@
-from domino.aisystems._util import build_eval_result_tag
+from domino.aisystems._eval_tags import build_eval_result_tag
 
 def test_build_eval_result_tags():
         assert build_eval_result_tag('my_metric', '1') ==  'domino.prog.metric.my_metric', 'numbers should be metrics'
